Implemented JWT Revocation, changed some names of the db attributes.

+++ app/routes/users.py
@@ -1,9 +1,12 @@
 from flask import Blueprint, request, jsonify
 from app.extensions import db
 from app.models.user import User
+from app.models.token_blocklist import TokenBlocklist
 from app.schemas.user_schema import UserSchema
 from app.utils.security import hash_password, role_required
-from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt, get_jti
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
+from datetime import datetime
+
 # Status codes are critical:
 
 # 200 OK → success
@@ -78,12 +81,36 @@ def update_user(user_id: int):
         
         user.username = data["username"]
         user.email = data["email"]
-        user.password_hash = hash_password(data["password_hash"])
+        user.password = hash_password(data["password"])
         user.role  = data["role"]
         db.session.commit()                 # Commit changes to db
         return user_schema.jsonify(user), 200
     
     return jsonify(message = "Access Denied"), 403
+
+# It is only logging out the current token, but it works. 
+@user_bp.route("/logout", methods=["POST"])
+@jwt_required()
+@role_required(["user", "admin"])
+def logout():
+    current_user_id = int(get_jwt_identity())
+    jti = get_jwt()["jti"]
+
+    # If user exists
+    user = User.query.get_or_404(current_user_id)
+    
+    # Get all active tokens for the user, both access and refresh
+    tokens = TokenBlocklist.query.filter_by(user_id = current_user_id, revoked_at=None).all()
+    if tokens:
+        for token in tokens:
+            token.revoked_at = datetime.now()
+    else:
+        # Edge case, JWT was not saved
+        token = TokenBlocklist(jti = jti, user_id = current_user_id, token_type = get_jwt().get("type", "access"), revoked_at = datetime.now())
+        db.session.add(token)
+        
+    db.session.commit()
+    return jsonify(message = "User succesfully logged out!"), 200
 
 
 # DELETE /<user_id> --> Delete a user    
@@ -91,8 +118,23 @@ def update_user(user_id: int):
 @jwt_required()
 @role_required(["admin"])
 def remove_user(user_id: int):   
+    claims = get_jwt()
+    
+    if claims["role"] != "admin":
+        return jsonify(message = "Access Denied"), 403
+    
     user = User.query.get_or_404(user_id)
-    db.session.delete(user)
-    db.session.commit()
-    return jsonify(message = ""), 204   # Return nothing
+    
+    try:    # Transaction Block
+        # Before deletion, mark revoked_at time for all active tokens
+        TokenBlocklist.query.filter_by(user_id = user_id, revoked_at = None).update({"revoked_at": datetime.now()}, synchronize_session = False)
+            
+        # Success --> Transaction Committed
+        db.session.delete(user)        
+        db.session.commit()
+        return jsonify(message = ""), 204 
+    except Exception as e:
+        # Failure --> Transaction Aborted, rollback automatically, but just to be sure we call it
+        db.session.rollback()
+        return jsonify(message = f"Error deleting user: {str(e)}"), 500
